Check whether the file name of a metadata file matches the model_id or combination of team_abbr and model_abbr specified within the metadata file

Check whether the file name of a metadata file matches the model_id or combination of team_abbr and model_abbr specified within the metadata file

Usage

check_metadata_file_name(file_path, hub_path = ".")

Arguments

file_path

character string. Path to the file being validated relative to the hub's model-metadata directory.

hub_path

Either a character string path to a local Modeling Hub directory or an object of class <SubTreeFileSystem> created using functions s3_bucket() or gs_bucket() by providing a string S3 or GCS bucket name or path to a Modeling Hub directory stored in the cloud. For more details consult the Using cloud storage (S3, GCS) in the arrow package. The hub must be fully configured with valid admin.json and tasks.json files within the hub-config directory.

Value

Depending on whether validation has succeeded, one of:

<message/check_success> condition class object.

<error/check_failure> condition class object.

Returned object also inherits from subclass <hub_check>.